+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to MiniJinja are documented here.
 
 ## 1.0.6
 
+- Fixed iso datetime formatting not handling negative offsets correctly.  #327
 - Re-report `Value` directly from the crate root for convenience.
 - It's now possible to `deserialize` from a `Value`.  Additionally the
   `ViaDeserialize<T>` argument type was added to support value conversions